fix(npm): add files prop to root package.json
This adds a files prop to the root package.json with an explicit list of all
of the files that should be included in the packages. This fixes a bug where
some distribution assets weren't getting included in the package while a lot
of non-essential development assets were. At the same time, this removes some
`src` directories from the packages which some consumers might have been using.